style of input
