+++ src/style/components/_CurrentUserProfile.scss
@@ -66,10 +66,10 @@
       z-index: 1;
 
       .image-container {
+        position: relative;
         .edit-image-overlay {
-          background-color: rgba(0, 0, 0, 0.486);
+          background-color: transparent;
           position: absolute;
-          visibility: hidden;
           z-index: 10;
           height: 10rem;
           width: 10rem;
@@ -84,12 +84,6 @@
           width: 10rem;
           height: auto;
           margin-bottom: 1rem;
-        }
-        &:hover {
-          cursor: pointer;
-          .edit-image-overlay {
-            visibility: visible;
-          }
         }
       }
     }
