+++ src/style/components/_EditPhotoModal.scss
@@ -1,0 +1,34 @@
+@use "../variables" as v;
+.photo-editor-modal {
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 100000;
+  background-color: v.$white;
+  text-align: center;
+  width: 10rem;
+  height: 10rem;
+
+  form {
+    height: min-content;
+    text-align: center !important;
+    .inputfile {
+      width: 0.1px;
+      height: 0.1px;
+      opacity: 0;
+      overflow: hidden;
+      position: absolute;
+      z-index: -1;
+    }
+    label {
+      background-color: rgba(0, 0, 0, 0.212);
+      padding: 1rem;
+      border-radius: 20px;
+      font-size: 1rem;
+      width: 100%;
+      &:hover {
+        cursor: pointer;
+      }
+    }
+  }
+}
