+++ src/style/index.scss
@@ -15,6 +15,7 @@
 @use "./components/ConnectForm";
 @use "./components/PreviewUserCard";
 @use "./components/Tags";
+@use "./components/EditPhotoModal";
 .App {
   font-size: 17px; // change this to variably increase the size of all the child rem sizing
   min-height: 100vh;
